% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimpLin.r
\name{SimpLinR}
\alias{SimpLinR}
\title{SimpLinR: Wrapper function for SimpLinCpp}
\usage{
SimpLinR(x, y)
}
\arguments{
\item{x}{numeric vector; predictor variable}

\item{y}{numeric vector; response variable}
}
\description{
Under the hood, SimpLinCpp performs simple linear regression in C++. This function acts as a wrapper
to catch errors.
}
